event-timing/interactionid-press-key-as-input.html

<!DOCTYPE html>
<html>
<meta charset=utf-8 />
<meta name="timeout" content="long">
<title>Event Timing: interactionId-press-key-as-input.</title>
<textarea id='test'></textarea>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/testdriver.js></script>
<script src=/resources/testdriver-actions.js></script>
<script src=/resources/testdriver-vendor.js></script>
<script src=resources/event-timing-test-utils.js></script>

<script>
  let observedEntries = [];
  let map = new Map();

  function eventsForCheck(entry) {
    if (entry.name === 'keydown' || entry.name === 'keyup') {
      map.set(entry.name, entry.interactionId);
      return true;
    }
    return false;
  }

  async function press(element_id, key) {
    const element = document.getElementById(element_id);

    const clickHandler = (e) => {
      mainThreadBusy(200);
    };

    element.addEventListener("keydown", clickHandler);

    await test_driver.send_keys(element, key);
  }

  async_test(function (t) {
    assert_implements(window.PerformanceEventTiming, 'Event Timing is not supported.');

    new PerformanceObserver(t.step_func(entryList => {
      observedEntries = observedEntries.concat(entryList.getEntries().filter(eventsForCheck));

      if (observedEntries.length < 2)
        return;

      assert_greater_than(map.get('keydown'), 0, 'Should have a non-trivial interactionId');
      assert_greater_than(map.get('keyup'), 0, 'Should have a non-trivial interactionId');
      assert_equals(map.get('keydown'), map.get('keyup'), 'keyDown and KeyUp should have the same interactionId');
      assert_equals('t', document.getElementById('test').value);
      t.done();
    })).observe({ type: "event" });

    press('test', 't');
  }, "Event Timing: compare event timing interactionId for key press as input.");
</script>

</html>
